frontend: split create-account into container+component
Also remove the old github-login-component which was already replaced I
think.
